mod stackwalking -> mod process_minidump
